+++ app/components/dashboard/cms/index.tsx
@@ -71,7 +71,7 @@ const CMSIndex = ({ singular = 'item', plural = 'items', items, loading, store, 
               </Button>
               <Button
                 onClick={() => router.push(`/dashboard/${plural}/new?store=${store.documentId}`)}
-                disabled
+                disabled={plural !== 'stores'}
                 leftSection={<IconPlus size={16} />}
               >
                 New {singular}
